# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: آل ياسين لاكسسوار الموتال - نظام الكاشير
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: List all debts
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [customer, supplier]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, partial, paid]
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
    post:
      operationId: createDebt
      tags: [debts]
      summary: Create a debt record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtBody"
      responses:
        "201":
          description: Created debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}:
    get:
      operationId: getDebt
      tags: [debts]
      summary: Get debt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
        "404":
          description: Not found
    patch:
      operationId: updateDebt
      tags: [debts]
      summary: Update debt (payment)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDebtBody"
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}/payments:
    post:
      operationId: addDebtPayment
      tags: [debts]
      summary: Add payment to a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPaymentBody"
      responses:
        "201":
          description: Payment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /customers/debts:
    get:
      operationId: getCustomerDebts
      tags: [debts]
      summary: Get debts by customer name
      parameters:
        - name: customerName
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cashier]
        permissions:
          $ref: "#/components/schemas/Permissions"
        createdAt:
          type: string
      required:
        - id
        - username
        - name
        - role
        - permissions
        - createdAt
    Permissions:
      type: object
      properties:
        canViewDashboard:
          type: boolean
        canManageProducts:
          type: boolean
        canManageSuppliers:
          type: boolean
        canMakeSales:
          type: boolean
        canViewReports:
          type: boolean
        canManageUsers:
          type: boolean
        canManageDebts:
          type: boolean
        canViewPurchases:
          type: boolean
        canCreatePurchases:
          type: boolean
      required:
        - canViewDashboard
        - canManageProducts
        - canManageSuppliers
        - canMakeSales
        - canViewReports
        - canManageUsers
        - canManageDebts
        - canViewPurchases
        - canCreatePurchases
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        quantity:
          type: integer
        minStockLevel:
          type: integer
        unit:
          type: string
        isLowStock:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - costPrice
        - sellingPrice
        - quantity
        - minStockLevel
        - unit
        - isLowStock
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        barcode:
          type: string
        category:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        quantity:
          type: integer
        minStockLevel:
          type: integer
        unit:
          type: string
      required:
        - name
        - costPrice
        - sellingPrice
        - quantity
        - minStockLevel
        - unit
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        barcode:
          type: string
        category:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        quantity:
          type: integer
        minStockLevel:
          type: integer
        unit:
          type: string
...
    Debt:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [customer, supplier]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
          enum: [pending, partial, paid]
        saleId:
          type: ["integer", "null"]
        purchaseId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        payments:
          type: array
          items:
            $ref: "#/components/schemas/DebtPayment"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - type
        - totalAmount
        - paidAmount
        - remainingAmount
        - status
        - payments
        - createdAt
        - updatedAt
    DebtPayment:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        notes:
          type: ["string", "null"]
        paidAt:
          type: string
      required:
        - id
        - amount
        - paidAt
    CreateDebtBody:
      type: object
      properties:
        type:
          type: string
          enum: [customer, supplier]
        customerName:
          type: string
        customerPhone:
          type: string
        supplierId:
          type: integer
        totalAmount:
          type: number
        paidAmount:
          type: number
        notes:
          type: string
        saleId:
          type: integer
        purchaseId:
          type: integer
      required:
        - type
        - totalAmount
        - paidAmount
    UpdateDebtBody:
      type: object
      properties:
        notes:
          type: string
    AddPaymentBody:
      type: object
      properties:
        amount:
          type: number
        notes:
          type: string
      required:
        - amount
...