# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * آل ياسين لاكسسوار الموتال - نظام الكاشير
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  category?: string | null;
  costPrice: number;
  sellingPrice: number;
  quantity: number;
  minStockLevel: number;
  unit: string;
  isLowStock: boolean;
  createdAt: string;
  updatedAt: string;
}