# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * آل ياسين لاكسسوار الموتال - نظام الكاشير
 * OpenAPI spec version: 0.1.0
 */
import type { Permissions } from "./permissions";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  permissions: Permissions;
  createdAt: string;
}